Update: Modularise Spinner component
