+++ src/components/adslotUi/TreePicker/TreePickerNodeExpanderComponent.jsx
@@ -1,6 +1,6 @@
 import React, { PropTypes } from 'react';
 import GridCell from 'components/alexandria/GridCell/component';
-import Spinner from 'components/alexandria/SpinnerComponent';
+import Spinner from 'components/alexandria/Spinner/component';
 
 const TreePickerNodeExpander = ({
   isLoading,
