+++ src/components/alexandria/Spinner/styles.scss
@@ -1,7 +1,7 @@
-@import '../color';
-@import '../font-weight';
-@import '../font-size';
-@import '../variable';
+@import 'styles/color';
+@import 'styles/font-weight';
+@import 'styles/font-size';
+@import 'styles/variable';
 
 .spinner-component {
   $color-spinner-light: $color-gray-light;
